Add function symbols to formula
This addition makes the algorithms and proofs quite a bit longer and
harder. But adding function symbols is essential for developing more
sophisticated expressions.

The approach I am now trying works as follows: a function symbol can
be used by providing a defining formula. The Realizes predicate
requires a value that materializes a given term in order to
evaluate a relation in the model. A value materializes a
function term if it satisfies the defining formula.